Initital commit of main.py
main.py is intended to be the main script of the project. As of the
first commit, it uses python-opencv to read webcam data, reads pictures
to show buttons on the screen, and overlays those images on the screen.
Some of these functions may be put into separate scripts later.